Refactor unit_test and function, add function dsPhanTuChung and test
